Support HYPH_LEVEL > 1
This could be useful for extending a base patterns file.

This requires generating a more advanced "dictionary" file based on a
previous patterns file with '*' (good hyphen ) and '.' (bad hyphen)
marks in addition to the '-' (missed hyphen) and '0' (don't care) marks.

In the future we could even automatically perform multiple patgen
stages. For now, we simply fail when bad or missed hyphens remain after
the first stage (and we adjust the parameters so that this doesn't
happen).

Also, as an additional check, generate a second dictionary file from the
generated patterns file and ensure that this file is identical to the
pattmp.X file produced by patgen.