Move everything out of the table-dev directory
- C code moved to liblouis
- test moved to tests
- everything else (Makefile, python, bash scripts) moved to tools

Also added a frontend/wrapper script for the Makefile. It is called
"lou_maketable" and adds some options and error handling and provides
a help text.

Some of the configuration is moved out of the Makefile into a
configuration file which must be passed as an option to lou_maketable.